Reorganize star database code
- Move StarDBBuilder to separate files
- Move cross-index catalogs into StarNameDatabase
- Use indices instead of pointers in catalogNumberIndex
