+++ src/celengine/starname.h
@@ -12,6 +12,8 @@
 
 #pragma once
 
+#include <array>
+#include <cstddef>
 #include <cstdint>
 #include <iosfwd>
 #include <memory>
@@ -19,10 +21,20 @@
 
 #include <celengine/name.h>
 
-class StarNameDatabase: private NameDatabase
+enum class StarCatalog : unsigned int
+{
+    HenryDraper,
+    SAO,
+    _CatalogCount,
+};
+
+class StarNameDatabase : private NameDatabase
 {
 public:
-    StarNameDatabase() {};
+    static constexpr AstroCatalog::IndexNumber TYC3_MULTIPLIER = 1000000000u;
+    static constexpr AstroCatalog::IndexNumber TYC2_MULTIPLIER = 10000u;
+
+    StarNameDatabase() = default;
 
     using NameDatabase::add;
     using NameDatabase::erase;
@@ -35,21 +47,38 @@ public:
     // We don't want users to access the getCatalogMethodByName method on the
     // NameDatabase base class, so use private inheritance to enforce usage of
     // the below method:
-    std::uint32_t findCatalogNumberByName(std::string_view, bool i18n) const;
+    AstroCatalog::IndexNumber findCatalogNumberByName(std::string_view, bool i18n) const;
 
+    AstroCatalog::IndexNumber searchCrossIndexForCatalogNumber(StarCatalog, AstroCatalog::IndexNumber number) const;
+    AstroCatalog::IndexNumber crossIndex(StarCatalog, AstroCatalog::IndexNumber number) const;
+
+    bool loadCrossIndex(StarCatalog, std::istream&);
     static std::unique_ptr<StarNameDatabase> readNames(std::istream&);
 
 private:
-    std::uint32_t findFlamsteedOrVariable(std::string_view, std::string_view, bool) const;
-    std::uint32_t findBayer(std::string_view, std::string_view, bool) const;
-    std::uint32_t findBayerNoNumber(std::string_view,
-                                    std::string_view,
-                                    std::string_view,
-                                    bool) const;
-    std::uint32_t findBayerWithNumber(std::string_view,
-                                      unsigned int,
-                                      std::string_view,
-                                      std::string_view,
-                                      bool) const;
-    std::uint32_t findWithComponentSuffix(std::string_view, bool) const;
+    static constexpr auto NumCatalogs = static_cast<std::size_t>(StarCatalog::_CatalogCount);
+
+    struct CrossIndexEntry
+    {
+        AstroCatalog::IndexNumber catalogNumber;
+        AstroCatalog::IndexNumber celCatalogNumber;
+    };
+
+    using CrossIndex = std::vector<CrossIndexEntry>;
+
+    AstroCatalog::IndexNumber findByName(std::string_view, bool) const;
+    AstroCatalog::IndexNumber findFlamsteedOrVariable(std::string_view, std::string_view, bool) const;
+    AstroCatalog::IndexNumber findBayer(std::string_view, std::string_view, bool) const;
+    AstroCatalog::IndexNumber findBayerNoNumber(std::string_view,
+                                                std::string_view,
+                                                std::string_view,
+                                                bool) const;
+    AstroCatalog::IndexNumber findBayerWithNumber(std::string_view,
+                                                  unsigned int,
+                                                  std::string_view,
+                                                  std::string_view,
+                                                  bool) const;
+    AstroCatalog::IndexNumber findWithComponentSuffix(std::string_view, bool) const;
+
+    std::array<CrossIndex, NumCatalogs> crossIndices;
 };
